to achieve dynamic remounting of CSI storage after a sandbox instance resumes from hibernation
