+++ pkg/controller/sandbox/core/interface.go
@@ -26,6 +26,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	agentsv1alpha1 "github.com/openkruise/agents/api/v1alpha1"
+	"github.com/openkruise/agents/pkg/sandbox-manager/clients"
+	"github.com/openkruise/agents/pkg/sandbox-manager/infra/sandboxcr"
 	"github.com/openkruise/agents/pkg/utils/expectations"
 	"github.com/openkruise/agents/pkg/utils/inplaceupdate"
 )
@@ -59,9 +61,17 @@ type SandboxControl interface {
 	EnsureSandboxTerminated(ctx context.Context, args EnsureFuncArgs) error
 }
 
-func NewSandboxControl(c client.Client, recorder record.EventRecorder, rl *RateLimiter) map[string]SandboxControl {
+type SandboxControlArgs struct {
+	Client        client.Client
+	Recorder      record.EventRecorder
+	RateLimiter   *RateLimiter
+	SandboxClient *clients.ClientSet
+	Cache         *sandboxcr.Cache
+}
+
+func NewSandboxControl(args SandboxControlArgs) map[string]SandboxControl {
 	controls := map[string]SandboxControl{}
-	controls[CommonControlName] = NewCommonControl(c, recorder, rl)
+	controls[CommonControlName] = NewCommonControl(args)
 	return controls
 }
